Test task
Plugin code is in [!etcetera-test-plugin] directory.

Note 1
Ajax filtering by ACF fields works, but there are no paginations in the results.
I think, this is a solvable task, but I have not met with it before and now there wasn't enough time to implement it

Note 2
I'm not sure if you want to use "flats" and their features in the real estate object as a repeater field. It is not available in the free version of ASF.

For the moment, the task is done on the principle of "one house - one apartment".